pyproject.toml
[tool.poetry]
name = "steampulse"
version = "0.1.0"
description = "AI-powered Steam review analytics for indie game developers"
authors = ["SteamPulse <hello@steampulse.gg>"]
readme = "README.md"
packages = [{include = "steampulse"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.135.1"
uvicorn = {extras = ["standard"], version = "^0.41.0"}
httpx = "^0.28.1"
anthropic = "^0.84.0"
jinja2 = "^3.1.6"
resend = "^2.23.0"
python-dotenv = "^1.2.2"
rich = "^14.3.3"
psycopg2-binary = "^2.9.11"
python-multipart = "^0.0.22"

[tool.poetry.group.crawler]
optional = true

[tool.poetry.group.crawler.dependencies]
boto3 = "^1.34.0"

[tool.poetry.group.dev.dependencies]
pytest = "^9.0.2"
pytest-asyncio = "^1.3.0"
httpx = "^0.28.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
steampulse = "steampulse.main:main"
